+++ src/RestApi/MapRestController.php
@@ -47,6 +47,9 @@ class MapRestController extends AbstractFOSRestController
 		if (in_array('fairteiler', $types)) {
 			$markers['fairteiler'] = $this->mapGateway->getFoodSharePointMarkers();
 		}
+		if (in_array('communities', $types)) {
+			$markers['communities'] = $this->mapGateway->getCommunityMarkers();
+		}
 		if (in_array('betriebe', $types)) {
 			if (!$this->session->id()) {
 				throw new UnauthorizedHttpException('Not logged in.');